leaderboard added and did some fixes

+++ src/cache.ts
@@ -1,35 +1,45 @@
+import {browser} from '$app/environment';
+
 export default class Cache {
 
     /**
      * get item from cache
-     * @param item 
-     * @returns 
+     * @param item
+     * @returns
      */
     static getCacheItem(item: string): any {
-        return window.localStorage.getItem(item);
+        if (browser) {
+            return window.localStorage.getItem(item);
+        }
     }
 
     /**
      * set item in cache
-     * @param item 
-     * @param value 
+     * @param item
+     * @param value
      */
     static setCacheItem(item: string, value: any) {
-        window.localStorage.setItem(item, JSON.stringify(value));
+        if (browser) {
+            window.localStorage.setItem(item, JSON.stringify(value));
+        }
     }
 
     /**
      * remove item from cache
-     * @param item 
+     * @param item
      */
     static removeCacheItem(item: string) {
-        window.localStorage.removeItem(item);
+        if (browser) {
+            window.localStorage.removeItem(item);
+        }
     }
 
     /**
      * remove all items from cache
      */
     static clearOnLogout() {
-        window.localStorage.clear();
+        if (browser) {
+            window.localStorage.clear();
+        }
     }
 }
